+++ webpack.config.js
@@ -2,8 +2,6 @@ var webpack = require('webpack');
 
 module.exports = {
 
-  entry: './modules/BufferedStream.js',
-
   output: {
     library: 'BufferedStream',
     libraryTarget: 'var'